Updated: polls reload on interval
